Added constraints, type money was replaced by a numeric, cosmetic changes

+++ tables/orders.sql
@@ -1,10 +1,10 @@
 create table shop.orders
 (
-	id serial
-		constraint orders_pk
+	id		serial
+			constraint orders_pk
 			primary key,
-	shop_id int,
-	total_amount money,
-	date timestamp,
-	employee_id int
+	shop_id	int,
+	total_amount	numeric(12,2),
+	date		timestamp,
+	employee_id	int
 );
